REFACTOR - Question 삭제기능 객체 안으로 넣음
