+++ README.md
@@ -1,3 +1,3 @@
 # 질문 삭제하기 기능 리팩토링
 [X] Answer 삭제 기능 리팩토링 - 답변자 확인하는 로직 & 삭제 기능을 객체 내부로 이동
-[ ] Question 삭제 기능 리팩토링 - 글쓴이 확인하는 로직 & 삭제 기능을 객체 내부로 이동
+[X] Question 삭제 기능 리팩토링 - 글쓴이 확인하는 로직 & 삭제 기능을 객체 내부로 이동
