+++ src/main/java/qna/domain/Answer.java
@@ -44,7 +44,7 @@ public class Answer extends AbstractEntity {
         this.contents = contents;
     }
 
-    public void delete(User writer) throws Exception {
+    public void delete(User writer) throws CannotDeleteException {
         if (!isOwner(writer)) throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
         setDeleted(true);
     }
