+++ src/main/java/qna/domain/Question.java
@@ -1,6 +1,7 @@
 package qna.domain;
 
 import org.hibernate.annotations.Where;
+import qna.CannotDeleteException;
 
 import javax.persistence.*;
 import java.util.ArrayList;
@@ -37,6 +38,15 @@ public class Question extends AbstractEntity {
         super(id);
         this.title = title;
         this.contents = contents;
+    }
+
+    public void delete(User writer) throws CannotDeleteException {
+        if (!isOwner(writer)) throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
+
+        for (Answer answer : answers) {
+            answer.delete(writer);
+        }
+        setDeleted(true);
     }
 
     public String getTitle() {
